requirements.txt
geopandas
rasterio
shapely
folium
ultralytics
sahi
boto3
huggingface_hub
matplotlib
numpy
psutil
Pillow
tqdm
torch
botocore
streamlit
plotly
